requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
python-dotenv==1.0.1
pydantic==2.10.0
slowapi==0.1.9
prometheus-client==0.21.0
pytest==8.3.0
httpx==0.27.2